refactor(core/domain): split alert inbox and listener
Move the HTTP listener into alert_listener.py and keep inbox.py focused
on the queue model and session accessors. Update the REPL controller and
tests to import listener symbols from the new module.

Part of #3775 (commit 6/11 on #3776).
