+++ core/domain/alerts/__init__.py
@@ -3,6 +3,7 @@
 - ``alert_source.py``  — resolve alert vendor, map to tool sources, relevance scoring
 - ``extraction.py``    — deterministic field extraction for the extract_alert stage
 - ``normalization.py`` — canonical OpenSRE alert payload shape
-- ``inbox.py``         — in-process alert queue and local HTTP listener
+- ``inbox.py``         — in-process alert queue
+- ``alert_listener.py`` — local HTTP listener for pushed alerts
 - ``tool_planning.py`` — score and rank investigation tools for an alert
 """
